+++ packages/chip/src/lib/StatusChip.tsx
@@ -1,0 +1,24 @@
+import { Chip as MuiChip, ChipProps as MuiChipProps } from '@mui/material';
+
+export type StatusChipProps = {
+  /** The color of the component.
+   * @default default */
+  color?: 'default' | 'primary' | 'secondary' | 'error' | 'info' | 'success' | 'warning';
+} & Omit<
+  MuiChipProps,
+  | 'avatar'
+  | 'children'
+  | 'color'
+  | 'variant'
+  | 'skipFocusWhenDisabled'
+  | 'disabled'
+  | 'size'
+  | 'icon'
+  | 'clickable'
+  | 'deleteIcon'
+  | 'onDelete'
+>;
+
+export const StatusChip = ({ color = 'default', ...rest }: StatusChipProps): JSX.Element => {
+  return <MuiChip color={color} {...rest} size="small" />;
+};