+++ packages/typography/CHANGELOG.md
@@ -6,7 +6,6 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ## 0.1.0 (2023-08-29)
 
-
 ### Features
 
-* **mui-typography:** add typography package ([4353d16](https://github.com/Availity/element/commit/4353d162749794063c816afb0bad06a0ae96105f))
+- **mui-typography:** add typography package ([4353d16](https://github.com/Availity/element/commit/4353d162749794063c816afb0bad06a0ae96105f))